pyproject.toml
[tool.ruff]
line-length = 120
target-version = 'py312'

[tool.poetry]
name = "redbox"
version = "0.5.0"
description = "RAG application that uses GenAI to chat with and summarise civil service documents"
authors = ["DBT Redbox <redbox@businessandtrade.gov.uk>", "i.AI <i-dot-ai-enquiries@cabinetoffice.gov.uk>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
django = "5.2.8"
django-libsass = "^0.9"
django-magic-link = "^1.0.0"
django-staff-sso-client = "^4.3.0"
dbt-copilot-python = "^2.0.0"
uwotm8 = "^0.1.6"
detect-secrets = "^1.5.0"
ddtrace = "^3.18.1"
django-log-formatter-asim = "^1.1.0"
django-requestlogs = "^0.8"
jinja2 = "3.1.6"
openpyxl = "^3.1.5"
coverage = "^7.11.1"
pydantic = "^2.7.1"
elasticsearch = "^8.15.0"
langchain-community = "0.3.27"
langchain = "^0.3.4"
langchain_openai = ">0.1.21"
boto3 = "^1.34.160"
pydantic-settings = "^2.3.4"
langchain-elasticsearch = ">0.2.2"
pytest-dotenv = "^0.5.2"
kneed = "^0.8.5"
langgraph = "^0.2.76"
langchain-aws = "^0.2.3"
wikipedia = "^1.4.0"
whitenoise = "^6.11.0"
watchdog = {extras = ["watchmedo"], version = "^4.0.2"}
opensearch-py = "^2.7.1"
scikit-learn = "^1.5.2"
django-environ = "^0.12.0"
django-use-email-as-username = "^1.4.0"
markdown-it-py = "^3.0.0"
psycopg2-binary = "^2.9.11"
pyyaml = "^6.0.3"
pytz = "^2024.1"
django-permissions-policy = "^4.28.0"
django-csp = "^3.7"
django-single-session = "^0.2.0"
sentry-sdk = {extras = ["django"], version = "^2.43.0"}
django-storages = "^1.14.2"
mohawk = "^1.1.0"
django-waffle = "^4.2.0"
langchain-mcp-adapters = "^0.0.11"
pymupdf = "^1.26.0"
pytest = "^8.4.2"
urllib3 = "2.5.0"
requests = "^2.32.5"
python-dotenv = "^1.2.1"
yarl = "^1.22.0"
humanize = "^4.14.0"
channels = {extras = ["daphne"], version = "^4.3.1"}
django-gov-notify = "^0.6.0"
django-import-export = "^4.3"
django-q2 = "^1.6.2"
blessed = "^1.23.0"
croniter = "^3.0.3"
djangorestframework = "^3.16.1"
django-plotly-dash = "^2.5.0"
django-adminplus = "^0.6"
pandas = "^2.3.3"
langchain-text-splitters = "0.3.9"
starlette = "0.49.1"
dj-database-url = "^2.3.0"
django-chunk-upload-handlers = "^0.0.15"
unoserver = "^3.4"
python-slugify = "^8.0.4"


[tool.poetry.group.dev.dependencies]
bumpversion = "^0.6.0"
pre-commit = "^3.6.2"
freezegun = "^1.5.5"
django-types = "^0.22.0"

pytest = "^8.4.2"
pytest-django = "^4.8.0"
pytest-env = "^1.2.0"
pytest-mock = "^3.15.1"
pytest-cov = "^5.0.0"
pytest-dotenv = "^0.5.2"
pytest-asyncio = "^1.2.0"
pylint = "^3.2.6"
ruff = "^0.14.5"
bandit = "^1.8.6"
detect-secrets = "^1.5.0"
types-python-dateutil = "^2.9.0.20240315"
types-markdown = "^3.10.0.20251106"
types-pytz = "^2024.1.0.20240203"
boto3-stubs = {extras = ["essential"], version = "^1.40.64"}
moto = {extras = ["s3"], version = "^5.1.16"}
playwright = "^1.55"
pytest-playwright = "^0.7"
axe-playwright-python = "^0.1"
websockets = "^12.0"
httpx = "^0.28.1"
minio = "^7.2.18"
yarl = "^1.22.0"
ipykernel = "^6.29.5"
jsonlines = "^4.0.0"
deepeval = "^1.0.3"
requests-mock = "^1.12.1"
django-test-migrations = "^1.3.0"
beautifulsoup4 = "^4.14"
gevent = "^24.10.2"
pandas = "^2.2.2"
dj-notebook = "^0.7.0"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocstrings = {extras = ["python"], version = "^0.30.1"}
mkdocs-material = "^9.6.23"
mkdocs-mermaid2-plugin = "^1.2.3"
markdown-mdantic = "^2.1.0"
pygments = "^2.18.0"
mkapi = "^2.1.0"
griffe-typingdoc = "^0.3.0"
mkdocs-git-revision-date-localized-plugin = "^1.5.0"
mkdocs-glightbox = "^0.5.2"
mkdocs-minify-plugin = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.format]
exclude = ["django_app/redbox_app/redbox_core/migrations/*.py"]

[tool.ruff.lint]
# Dribble in those we have commented out, asterisked ones first
select = [
    "A",
    "B",
    "C",
    "E",
    "F",
    "G",
    "I",
    "N",
    "Q",
    "S",
    "T",
    "W",
    "ARG",
    "ASYNC",
    "BLE",
    "COM",
    "DJ",
    "DTZ",
    "EM",
    "EXE",
    "FURB",
    "ICN",
    "INT",
    "ISC",
    "LOG",
    "NPY",
    "PD",
    "PGH",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "PYI",
    "RET",
    "RSE",
    "RUF",
    "SIM",
    "SLF",
    "SLOT",
    "TCH",
    "TD",
    "TID",
    "TRY",
    "UP",
    "YTT",
]
exclude = ["django_app/redbox_app/redbox_core/migrations/*.py", "out/*"]
ignore = ["COM812", "DJ001", "RET505", "RET508", "PLR0913", "ISC001"]

[tool.ruff.lint.isort]
known-first-party = ["redbox_app", "redbox", "core_api"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S106", "PLR0913", "PLR0915", "PLR2004", "TD003", "S311"]
"*/tests/*" = ["S101", "S106", "PLR0913", "PLR0915", "PLR2004", "TD003"]
"django_app/tests_playwright/*" = ["S101", "S106", "PLR0913", "PLR2004", "TD003"]
"worker/*" = ["B008"]
"redbox/*" = ["TCH003"]
"redbox/tests/*" = ["ARG001"]
"*/admin.py" = ["RUF012"]
"*/models.py" = ["RUF012"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "redbox_app.settings"
testpaths = "tests"
norecursedirs = "django_app/tests/test_ai.py"
env_override_existing_values = 1
env_files = [
    "django_app/tests/.env.test",
    ".env"
]
